MB-50423: Allow PagingVisitor to remove deleted item metadata
If a request needs to read a deleted value's body (e.g., to read system
xattrs), the delete is bgfetched back into the HashTable.

After this, under value eviction only the value could be evicted -
there was no mechanism to remove the deleted value metadata, other than
overwriting the document or recreating the vbucket. It could remain in
memory long after the tombstone has been purged from disk.

To resolve this, allow the PagingVisitor to remove deleted StoredValues
from the HT, even under value eviction.

This does not fundamentally change how deletes are handled; once a
delete is persisted it is removed from the HT entirely anyway, and only
bgfetched if needed.

Change-Id: I16f9c4c9455ffec430dbb79afe088ba46ad9de20
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169245
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>