Fix logging of SeqIterator in MagmaKVStore::initBySeqnoScanContext()
Fix the logging of the SeqIterator pointer so that it logs the address
of the SeqIterator ptr rather than the ptr of DomainAwareSeqIterator.

This patch also moves the to string code for ValueFilter &
DocumentFilter into a stream operator so they can be use directly with
fmt thought our code base.

Change-Id: I2786d05a999f55da97944cb20c3c8c7245b4b5d6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169243
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>