MB-37858: Introduce the seqno_persistence_timeout config param
That's the max duration (in seconds) after which a SeqnoPersistence
request is temp-failed.
Param's value validation is achieved within the EPConfig fixture.
The new param is dynamic and replaces the previously hard-coded value.
Default/min/max values for the new param keep the legacy values, so
there's no behaviour change with this patch.

Change-Id: I2cb2856c2ef7c96fd6e31a81b8adf7e485427a95
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169252
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>