MB-50389: Remove dirtyAge & dirtyAgeHighWat
EPStats::dirtyAge records the 'dirty age' of the last mutation to be
flushed - i.e. the duration from when the item was queued into the
CheckpointManger until it was processed by the flusher. In a live
system the duration of the last mutation is pretty meaningless - it's
basically just a single random sample. We already record a histogtram
of the dirty age for _all_ mutations flushed to disk, which is much
more useful.

Similary, EPStats::dirtyAgeHighWat records the maxmimum value for
dirtyAge - which the histogram also records for us.

Remove these two statistics as they just duplicate information
available elsewhere.

Change-Id: Icaaf6c86c89dec5612921b49ee0846b4b328b260
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169326
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>