MB-50555: Ensure max compaction concurrency is not exceeded if VB locked
During investigation of MB-50389, it was observed that more Compaction
tasks were concurrently running than expected. For example, when
configuring a maximum of 2 Compaction tasks to run concurrently, we
intemittently saw three Compaction tasks executing concurrently.

The issue is how we handle Compaction tasks which cannot run because
the vBucket is currently locked (e.g. by the Flusher). These tasks
should re-scheduled to try compacting again - and indeed we do see
this behaviour. However, an additional CompactTask (assuming there are
some waiting) can also be woken - causing the limit to be exceeded.

The issue is that the code to wake a queued Compaction task at the end
of the current Compaction does not check if the current compaction is
actually going to be re-scheduled.

Change-Id: Ic2e25254de5f6bca612d4f57b24b10e0764b8cb3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169347
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>