MB-50564: Only use "CA file" as the verify locations
Memcached behaved differently than the other services
in the way that we returned the full certificate chain
as trusted CA list and would also allow for clients
to connect if the certificate was signed by one of
the certificates in the chain.

This patch change the logic to be identical to
the other services so that only the certificates in
the "CA file" in the configuration gets reported
back to the client during the handshake, and only
the "CA file" is used for verification.

Change-Id: I0bda90c2d0ee39137352731db1668d0f55b6e5cd
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169459
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>