MB-50331: Inhibit background compaction during rollback for magma [2/2]
magma requires that no background compactions run during a rollback so
has code to marshal them and prevent them from running until a rollback
has finished. As this code is in magma::Rollback it is called while we
hold the vBucket state lock which we acquired in KVBucket::rollback.
Compactions calling back into kv_engine with expiries need to take
the vBucket state lock while they perform the operation so we can
deadlock if we rollback while a background compaction runs if it
has to expire an item.

Fix this issue by preventing background compactions from running
during rollback before we take the vBucket state lock, and allowing
them to run again after.

Change-Id: Id7cd18ed855fd8eb3bd62fba36d37d2dc060d81d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169479
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>