MB-50492: Nexus don't call getCollectionStats() too early
Currently NexusKVStore::generateSecondaryVBManifest() calls
KVStore::getCollectionStats() which access the local store to get
collection stats from disk. This is however, can cause very noisy logs
if the secondary KVStore hasn't had it's on disk collections haven't
written yet e.g. if generateSecondaryVBManifest() is being called for
the fist commit() of manifest items to disk.

To fix this, call getPersistedVBucketState() to see if we've written the
vbucket state to disk yet. Also ensure that when returning the
secondaryManifest, that the collections write lock isn't held.

Change-Id: If1c352bc9b32f2bc709848771b7967f142de0f45
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169485
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>