MB-49469: Introduce CheckpointManager::Counter::memFreedByExpel
This patch introduces the new CheckpointManager::Counter type.
That's for CM-level stats that need to reflect their updates into
EPStats.

The new type is also used here for the new CM::memFreedByExpel, which
account the memory released by ItemExpel by a specific CM instance.

The new quantity will be used in a follow-up for computing and exposing
new vb_{state} aggregates for the same quantity.

Change-Id: I2846833f9a7fec99966a05f9ed0190f045e23f59
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169495
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>