MB-50589: Move lastReadSeqno for deletes during delete scan
Currently when scanning with the NO_DELETES option (during warmup)
any deleted item does not invoked the supplied callback and does
not move the lastReadSeqno. Warmup also pauses if the scan takes more
than 10 milliseconds. Should we encounter a case where we have n
deleted items, followed by an alive item (which calls the callback
which pauses the scan) and the scan of the n deleted items takes
more than 10 milliseconds then we'll scan again from the start of
the deleted items. This can make indefinitely hang if it encounters
a range of deleted items that it cannot scan in under 10 milliseconds.

Fix this issue by moving the lastReadSeqno on deleted items if we
are scanning with the NO_DELETES option so that we do not reset our
scan progress if the first non-deleted item paused the scan.

Change-Id: I2099b8c3e7ba70086b558e9ca4f52ec28fd5065b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169497
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>