Fix Nexus implicit compaction tests
Magma only allows rollback if an item is not in a rollbackable
checkpoint (i.e. the latest n checkpoints). Our tests were written
before this change was made but they did not break when we made the
change even though they are not purging an item from the magma
KVStore that they were intended to. Correct this and add a stats
checks to determine that the secondary KVStore is in the appropriate
state (we can't just return get the value as we return primary values)

Change-Id: I74e4120c5be870e63e9a99a167cdebcffb7b2c4b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169604
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>