MB-50389: Don't use CompactTask when compacting in unit tests
A number of unit tests want to trigger compaction in specific places,
often interleaved with DCP backfill.

Currently this is done by scheduling a compaction task using the same
API as the real server does; then running the given task and waiting
for it to complete.

A subsequent patch will move CompactVBucketTask from Writer to AuxIO
thread; this makes running compaction in this way problematic as we
sometimes already have AuxIO tasks (Backfill) waiting to run, and hence
attempting to schedule a CompactVBucketTask but then run the "next"
task will actually run backfill.

As such, modify the unit test functions which are used to perform
compaction without the intermediate Task, instead just call the sync
method EPEngine::doCompact().

Also move where we apply config.retain_erroneous_tombstones to
compaction from CompactVBucketTask to doCompact() so that will occur
even if the function is called directly.

Change-Id: Id6d463880f223dbeec291f627c45f14eb6b5686c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169639
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>