MB-50389: Move CompactVBucketTask from Writer to AuxIO thread pool
+Summary+

This patch moves CompactVBucketTask from the Writer thread pool (where
the FlusherTask also runs) to the AuxIO thread pool.

This significantly reduces the tail latency for FlusherTasks as they
are no longer contending on the same thread pool with
CompactVBucketTask - which are typically 2 orders of magnitude slower
than FlusherTask.

+Background+

CompactVBucketTask has long been a Writer task, running on the same
Writer thread pool as FlusherTask. Prior to SyncWrites (v6.5.0) this
was a sensible choice - both were nominally background tasks which we
did not care that much about their latency; document mutations were
acknowledged to the client as soon as they were updated in memory and
were persisted in the background.

Starting with SyncWrites in 6.5.0, this changed - we now needed the
FlusherTaskr (for SyncWrites at least) to complete as quickly as possible
to minimise SyncWrite latency / maximise throughput.

To this end we we made a number of changes to flushing & compaction as
part of v6.5.0 and subsequent releases. Most significant were:

1. In 6.5.0; increase the default number of shards from 4 to #CPUs,
   and to add a 'Disk I/O Optimized' Writer thread setting which
   creates one Writer thread per CPU core. Together these allow us to
   increase flusher concurrency from 4 to much larger values;
   typically #CPUs. This reduces the time that a given vBucket is
   waiting to be flushed to disk (assumming multiple vBuckets are
   outstanding) and hence reduce SyncWrite latency.

2. In 7.0.0; allow Couchstore compaction to run concurrently with
   flushing - previously when a given vBucket was being compacted,
   that vBucket could not be flushed. For large vBuckets this could
   result in mulit-second pauses in Flushing (and hence SyncWrites).
   By allowing compaction to run at the same time for the majority of
   its runtime, only having to take exclusive access of the vbucket at
   the end when finishing up, we could reduce the tail latency of
   flushing when compaction was running significantly.

3. In 7.1.0; allow multiple vBuckets to be compacted
   concurrently. This is a feature which actually existed for a long
   time in kv-engine, however it was found to result in signifcant
   latency spikes for background fetches, so the number of concurrent
   compactors was reduced to 1 via MB-18426 in v4.5.1.  Having only a
   single compactor was found to limit the write throughput of large
   datasets / at high write rates; so concurrent compaction was
   re-enabled in 7.1.0 - much of the original latency impact had since
   been addressed so it was broadly safe to do so - see MB-49858.

(1) and (2) have improved FlusherTask significantly, but the
introducion of parallel compactors; necessary for keeping up with
larger workloads has brought the focus back to the impact that
Compaction has on flushing.

+Details+

Analysis of SyncWrite and persistTo p99.9 latency shows that
signifcant spikes in latency occur during Compaction. There are
broadly two reasons for these spikes:

a) Contention on Writer thread pool between FlusherTask and
   CompactVBucketTask means Flusher tasks can be delayed in being run.

b) Contention on the underlying disk subsystem means that it can take
   longer for write()s from FlusherTasks to be synced to the medium.

While both are important, (a) seems to have the larger contribution to
tail latency. In the case of 4 Writer threads, even running a single
Compaction task (while FlusherTasks are scheduled to run) effectively
removes 25% of the Writer thread capacity that the FlusherTasks
previously had exclusively for themselves.

Additionally, each Compactor task can easily take 2-3 orders of
magnitude more than a FlusherTask, so single Compactor task run can
increase the queue time of a blocked FlusherTask by 1000x of the mean
queue time. We also see situations where multiple Compactor tasks are
scheduled consecutively on the same thread, without scheduling any
FlusherTasks in between - further increasing the worsr-case queue time
for FlusherTasks.

While we could enhance how we prioritise different task types - trying
to space out Compactor tasks if FlusherTasks are waiting - the sheer
magnitude of the difference in runtime makes it very difficult to
completely address any tail latency issues caused by queueing. Even a
single Compactor task run potentially adds a very large latency to the
next FlusherTask.

Alternatively, we could look to move the CompactVBucketTask elsewhere
to a different thread pool. The FlusherTask is arguably the most
important task on the Writer pool; all other tasks (aside from
Compaction) are either short-running (similar duration to FlusherTask)
or are mutually exclusive with FlusherTask - such as Rollback. Having
long running task which is not mutually exclusive FlusherTask, and
whose latency is not important is a poor fit.

The AuxillaryIO thread pool manages a number of tasks with similar
properties as CompactVBucketTask - potentially long-running, not
latency-sensitive such as Backfill, AccessScanner, and
VBucketMemoryAndDiskDeletionTask. We have also recently increase the
default thread count of the AuxIO pool (MB-49786) to similar numbers
as the Writer pool, so there should be sufficient capacity to support
an additional task type.

Results of test "99.9th percentile PERSIST_TO latency (ms), 4 nodes, 1
bucket x 20M x 1KB, 10K ops/sec"
(kv_latency_mixed_20M_durability.test) for the following three
configurations:

     Build              Compactors  CompactVBucketTask  p99.9 latency (ms)

     7.1.0-1885         1           Writer pool         124
     7.1.0-2173         2           Writer pool         217
     7.1.0-14335 (toy)  2           AuxIO pool           64

i.e. p99.9 latency is 0.5x compared to the initial baseline, and 0.3x
of the latency compared to when compaction is on writer pool.

Change-Id: I428e37cc1b4a004648a63144a4ecbd4560d38425
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169640
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>