MB-50600: Use KVStoreRevision in prepareToDelete et. al.
Magma and couchstore vBuckets deal with revisioning slightly differently
if an item has not yet been written to a vBucket. Nexus needs to be able
to deal with that when we delete a vBucket revision as currently it tries
to delete the revision of the primary KVStore on the secondary. This may
fail if the KVStores revisions are not the same. Add a KVStoreRevision
class which wraps the uint64_t revision value to allow NexusKVStore to
use polymorphism to return a child type containing both revisions in a
future change.

Change-Id: I6ffe388da3744d8a439f3e989602ce08cb7ae17b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169745
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>