MB-50425: Add() of deleted item should fail if tombstone created during request
As part of adding support for allowing SDK Transactions to atomically
insert new documents ((MB-37374), the subdoc API was enhanced to allow
creating a document directly in the Deleted state - i.e. the operation
would only succeed if the docuemnt did not exist, and also a tombstone
did not exist.

Checking if a key already exists (alive or tombstone) was performed as
part of the standard subdoc "read-modify-write" execution model,
specifically in the 'read' step; however that failed to account for
the possibility that another concurrent client created a deleted
document sometime between the first client 'read' and 'write' steps.

In the case of a 'normal' Insert (of an alive document), this is not a
problem as the VBucket::add() method is the final arbiter and ensures
that no alive document exists (with the appropriate HashTable lock
held). However, VBucket::add() does not check what kind of document
(alive/deleted) is being inserted, and hence fails to perform the
appropriate check when inserting deleted documents.

The consequence of this is that concurrent attempts to Insert a
Deleted document via subdoc can result in more than one client
succeeding; as at the point they perform the 'read' phase; the
document doesn't exist (alive or tombstoned).

Address this by adding the missing checks in VBucket::add().

(cherry picked from commit 017aad8592d5e5765de5f64c98eec7235e8fbe46)

Change-Id: Ide367020f1a239cfe8797b09b2a59e878d7e300a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181091
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>