Remove nexus magma_couchstore variants from DurabilityWarmupTest
We are often seeing timeouts in this test suite and the
magma_couchstore variants add little value beyond the
couchstore_magma variants. Remove them from this test suite to speed
up runtimes and reduce the likelihood of a timeout. Other test suites
are still running the magma_couchstore variant as it exercises some
nexus code specific to that variant.

Change-Id: I63aabe1560e2306aef6ed9c877970624b06c7272
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169753
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>