MB-49512: Use min of AuxIO and Writer threads to calc compaction limit
Whilst it made sense to limit the concurrent compaction tasks by using
the number of writers threads when we running the tasks on the writer
pool it does not make sense when running them on the AuxIO pool. If we
had more writers than AuxIO threads then we could easily saturate the
AuxIO pool with the number of tasks that we allow to run concurrently.
This would negatively impact backfill performance. However, it's also
possible now to have more AuxIO threads than writer threads. Whilst we
could use the number of AuxIO threads for the calculation we could end
up with more concurrently running compaction tasks than flusher tasks
and that could negatively impact flush latencies if we saturate the
disk. As such, use the min of Writer and AuxIO threads in the
calculation to be more conservative with the number of compactions tasks
that we allow to run concurrently.

Change-Id: Ic30514c79ca57102b7c8bdac6a5e691abe236b96
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170089
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>