MB-50747: "logical-insert" of a tombstone leads to negative item count
Adding range checks on the item count detected that one unit test was
producing a negative item count when the backend is magma.

The following sequence of operations occurs.

seq:operation => expected item count

1:create collection c1 =>          1
2:insert key1 into c1 =>           2

flush

3:drop collection c1 =>            1
4:create collection c1 =>          2
5:insert a deleted key1 into c1 => 2
6:drop collection c1 =>            1
flush

Prior to this fix after this sequence when compaction runs the
final item count became -1

This occurs because seq:5 was treated as a plain delete and set
the magma `docCount` for that flush as -2 (ninserts:0 + ndeletes:2).
I.e. drop collection was -1 and the delete was -1.

The compaction then took the item count of 1 and subtracted -2.

The fix here skips the ndeletes count for the store at seq:5, the
'logically inserted tombstone' should have no affect on the item
count.

Change-Id: Ieae78eb9f61d2555650a46010845dfc909cc1309
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170110
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>