MB-50543: Ensure we don't send non-monotonic snapshot start
Problem:
MB-50543 demonstrated that in some situations the ActiveStream would
send a DCP Snapshot with a start seqno that was less than the previous
snapshot's start seqno, causing a monotonic violation. We also didn't
see this within KV-Engine testing as we don't have any monotonic checks
on snapshot start and end seqnos for our DCP producer or DCP consumer.
Further more this doesn't happen regularly and is more likely to happen
when streaming collections due to items being non visible.

Cause:
This was due to ActiveStream::nextSnapStart being used in a situation
where we had already send a snapshot with a snapEndSeqno grater than
it. This could happen as we would update ActiveStream::nextSnapStart
when processing new items from a checkpoint in
ActiveStream::processItems(), where we had items for multiple checks
points but not the checkpoint start for the first set of items. This
would mean that we would use the first checkpoints range to set
ActiveStream::nextSnapStart instead of it's checkpoints range.

Fix:
Instead of ActiveStream::nextSnapStart being option make it weakly
monotonic to ensure we ratchet the value. Also when we consider
modifying the sapStart in ActiveStream::snapshot() only do this if
nextSnapStart is greater than the last snapshot's snapEndSeqno.

Change-Id: I5488e1daade5a8e7ea10c8225d7cf15128ee7960
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170120
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>