MB-49512: Move logging and common executor code to GlobalTask::execute
Currently we don't log the bucket if a thread schedule or run is slow
which can sometimes be useful information. We don't have this
information because we don't set the thread local via a
BucketAllocationGuard before calling these functions. Correct this by
moving the code into the GlobalTask::execute function (as the code
is common for both ExecutorPool implementations) and under the
BucketAllocationGuard that we already take to execute the task (so that
we don't ping-pong between contexts/memory arenas).

Change-Id: I1e2f42dd3a02767e325ce84486fbedce28ac5161
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170153
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>