MB-50620 Fix initial disk snapshot check
In commit fd262fe8a we introduced an optimisation that let us issue
Insert operations intead of regular Upserts if a replica is receiving an
initial disk snapshot. This situation was detected by checking if the
received snapshot marker has startseqno == 0.

This condition is incorrect since if a backfill is interrupted and
resumed later, producer still sends a snapshot marker with
snap_start_seqno=0, but starts sending items (start_seqno) after the
consumer's highSeqno. These items are not eligible for Insert operations
as those items could already be present in the vbucket. To maintain
correct item count a lookup on set i.e. Upsert is required.

Fix this by detecting an initial disk snapshot on the basis of vbucket's
highSeqno. This is also how VBucket::setReceivingInitialDiskSnapshot()
is implemented.

Change-Id: I9fd87632fa70c3ce63d193470fd1a9207ac33875
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170171
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>