MB-50543: Make PassiveStream's seqnos monotonic
Ensure that the our consumer streams throw if they receive a seqno that
isn't weakly monotonic for current snapshot start and monotonic end
seqnos, and also the last received seqno. For cur_snapshot_start use the
default throw policy as this will mean we won't throw in production
builds, as it we're sent a non monotonic snapshot start seqno it's not
the end of the world. But for the snapshot end seqno and last_seqno,
could cause nasty data inconsistency issues.

This patch also fixes up test so they correctly drive passive streams by
not causing these values to be not monotonic.

Change-Id: I0702d2be100ff1601e59b8f9036c4c3d9c8005f9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170244
Tested-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>