MB-50543: Make Checkpoint::snapEndSeqno WeaklyMonotonic
This patch makes Checkpoint::snapEndSeqno WeaklyMonotonic to ensure we
never go backwards in history when trying to add modify a checkpoint's
range. We've chosen to use a WeaklyMonotonic semantic, as in some
cases we'll update the end seqno to the value it currently is.

This patch also add fixes to tests that incorrectly created checkpoints
or tried to add seqnos out of a checkpoints range.

Change-Id: I1de6a766d61b45fc88e481f6b76175ce5bdd876e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170285
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>