Refactor: Move JSON validation to FrontEndThread member
Previously we had the JSON validator as a member in the
FrontEndThread struct, and everyone who wanted to use
it could use it the way they wanted.

This patch refactor this to a member function:

    bool isValidJson(Cookie& cookie, std::string_view view);

The member function adds a trace span and updates the bucket
histogram for the validation time.

Call the same method from subdoc validation instead of
creating a new validator every time we need one.

Change-Id: I7e29ceecb6eee3abf2a1d6ee2c339f1c0b0d2206
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170345
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>