MB-50747: Compaction may skip dropped collection item count updates
MagmaKVStore drops collections in the implicit compaction as follows.

1. Compact and find all tables storing prepares
2. Iterate through the list of dropped collections
2.1 Compact and find all tables storing keys of that collection

During 2.1 the item count of that collection was decremented from the
vbucket total.

However a bug exists in that steps 1 and 2.1 will attempt to purge all
keys from the found table(s), this means 1 or 2.1 could also fully
purge a collection that we've yet to reach in step 2.

When this happens, the item count of that collection doesn't get
decremented from the total, and the vbucket item count now shows more
items than it stores.

To fix add an optional collection-ID to the compaction context. When
this optional stores a value, the compactionCore will only drop keys
of that collection.

Secondly move the prepare namespace purge to the last namespace
compacted so that it doesn't encounter keys from dropped collections
other than any remaining prepares not dropped.

Change-Id: Ia8841139f965d01a5de039e5df0eddc851ecd22d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170366
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>