MB-50647: Remove ConnMap::releaseLock
The releaseLock was originally designed to synchronise cookie release
and notification. Once released, it was no longer correct to notify the
associated connection.

Since https://review.couchbase.org/c/kv_engine/+/167564/ the cookie has
been managed in an RAII manner - if the ConnHandler still exists, the
cookie is still reserved.

As such, if a shared_ptr is held, there is no danger of the cookie
being concurrently released, so the lock is unnecessary.

Change-Id: I3e6e99c4b360bf4ad6175dd5688e7077f13f1a80
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170457
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>