MB-49512: Reset tasks on scheduler thread [2/2]
Currently the FollyExecutorPool uses a scheduler thread to deal with
any task scheduling, waking, or cancelling, and all work is done on
the CPUThreadPoolExecutor for some pool type (e.g. Writer/Reader).
One task currently run on the cpuPools is the cleaning up of task
pointers when tasks are cancelled. This is problematic during a bucket
shutdown (unregisterTaskable) as we currently must wait for task
pointers to be reset before we destruct the bucket. Should we have
contention on some cpuPool then bucket shutdown is blocked by the
running tasks to clean up the task pointers.

This is part 2 of 2 to allow the scheduler thread to perform the reset
of the task pointer rather than the cpuPool threads.

Change-Id: I620248be55a816811dbc8fe3f1e227eef988650f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170542
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>